% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_voteview}
\alias{read_voteview}
\title{Read Voteview data}
\usage{
read_voteview(chamber = c("both", "house", "senate"), congress = NULL,
  year = NULL, quiet = TRUE)
}
\arguments{
\item{chamber}{Specify which chamber of Congress–default is "both"}

\item{congress}{Specify which number of Congress. The default is the current
Congress.}

\item{year}{Optionally instead of specifying congress you may specify a year
or Date and the congress number will be inferred}

\item{quiet}{Logical indicating whether to silence column parsing messages}
}
\value{
A tibble of voteview information
}
\description{
Returns data frame of data provided by Voteview.com
}
